Consolidate line center pos logic
